+++ contracts/access/Roles.sol
@@ -3,9 +3,7 @@ pragma solidity ^0.4.24;
 
 /**
  * @title Roles
- * @author Francisco Giordano (@frangio)
  * @dev Library for managing addresses assigned to a Role.
- * See RBAC.sol for example usage.
  */
 library Roles {
   struct Role {
