+++ contracts/mocks/SignatureBouncerMock.sol
@@ -1,9 +1,10 @@
 pragma solidity ^0.4.24;
 
 import "../drafts/SignatureBouncer.sol";
+import "./SignerRoleMock.sol";
 
 
-contract SignatureBouncerMock is SignatureBouncer {
+contract SignatureBouncerMock is SignatureBouncer, SignerRoleMock {
   function checkValidSignature(address _address, bytes _signature)
     public
     view
