+++ contracts/token/ERC20/ERC20Mintable.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.4.24;
 
 import "./ERC20.sol";
-import "../../access/rbac/MinterRole.sol";
+import "../../access/roles/MinterRole.sol";
 
 
 /**
