+++ test/token/ERC20/ERC20Mintable.test.js
@@ -1,6 +1,6 @@
 const { shouldBehaveLikeERC20Mintable } = require('./ERC20Mintable.behavior');
 const ERC20MintableMock = artifacts.require('ERC20MintableMock');
-const { shouldBehaveLikePublicRole } = require('../../access/rbac/PublicRole.behavior');
+const { shouldBehaveLikePublicRole } = require('../../access/roles/PublicRole.behavior');
 
 contract('ERC20Mintable', function ([_, minter, otherMinter, ...otherAccounts]) {
   beforeEach(async function () {
